requirements.txt
anthropic>=0.49.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
lxml>=5.1.0
